+++ NoSimpleWayAPI[v2]/ext/includes/NoSimpleWayAPI[v2]/EDataCore.h
@@ -861,7 +861,7 @@ public:
 	static bool				if_text_is_float(std::string _text);
 
 	static int				safe_convert_string_to_number	(std::string _text, int _min,	int _max);
-	static int				safe_convert_string_to_float	(std::string _text, float _min, float _max);
+	static float			safe_convert_string_to_float	(std::string _text, float _min, float _max);
 
 	static std::string		string_array[512];
